pyproject.toml
[tool.poetry]
name = "planeks-test"
version = "0.1.0"
description = ""
authors = ["Kirill Shitikov <shitikovkirillt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.9"
drf-jwt = "^1.13"
drf-access-policy = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pre-commit = "^1.18"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
